Remove 'ExternalLink', create composable Link component with styling
