+++ app/Link.module.scss
@@ -1,0 +1,27 @@
+.link {
+  &:link,
+  &:visited {
+    color: var(--link);
+    display: inline-block;
+    text-decoration: none;
+    border-bottom: 1px solid var(--link);
+    padding: 2px;
+    transition: all 0.2s;
+  }
+
+  &:hover {
+    background-color: var(--link);
+    color: var(--white);
+    box-shadow: 0 1rem 2rem rgba(var(--black), 0.15);
+    transform: translateY(-2px);
+  }
+
+  &:active {
+    box-shadow: 0 0.5rem 1rem rgba(var(--black), 0.15);
+    transform: translateY(0);
+  }
+
+  &:has(+ button) {
+    margin-right: 0.35rem;
+  }
+}
